+++ batch/subsystem.go
@@ -1,19 +1,14 @@
 package batch
 
 import (
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"reflect"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/contribsys/faktory/client"
 	"github.com/contribsys/faktory/manager"
 	"github.com/contribsys/faktory/server"
-
 	"github.com/contribsys/faktory/util"
 )
 
@@ -48,7 +43,7 @@ func (b *BatchSubsystem) Start(s *server.Server) error {
 	if err := b.loadExistingBatches(); err != nil {
 		util.Warnf("loading existing batches: %v", err)
 	}
-	b.addCommands()
+	server.CommandSet["BATCH"] = b.batchCommand
 	b.addMiddleware()
 	util.Info("Loaded batching plugin")
 	return nil
@@ -62,148 +57,6 @@ func (b *BatchSubsystem) Name() string {
 // Reload does not do anything but is required by Faktory for subsystems
 func (b *BatchSubsystem) Reload(s *server.Server) error {
 	return nil
-}
-
-// Shutdown does not do anything but is required by Faktory for subystems
-func (b *BatchSubsystem) Shutdown(s *server.Server) error {
-	return nil
-}
-
-func (b *BatchSubsystem) addCommands() {
-	server.CommandSet["BATCH"] = func(c *server.Connection, s *server.Server, cmd string) {
-		util.Info(fmt.Sprintf("cmd: %s", cmd))
-		parts := strings.SplitN(cmd, " ", 3)[1:]
-		if len(parts) < 2 {
-			c.Error(cmd, errors.New("Invalid BATCH command"))
-			return
-		}
-
-		switch batchOperation := parts[0]; batchOperation {
-		case "NEW":
-			var batchRequest NewBatchRequest
-			if err := json.Unmarshal([]byte(parts[1]), &batchRequest); err != nil {
-				_ = c.Error(cmd, fmt.Errorf("Invalid JSON data: %v", err))
-				return
-			}
-
-			batchId := fmt.Sprintf("b-%s", util.RandomJid())
-
-			success := ""
-			if batchRequest.Success != nil {
-				successData, err := json.Marshal(batchRequest.Success)
-				if err != nil {
-					_ = c.Error(cmd, fmt.Errorf("Invalid Success job"))
-					return
-				}
-				success = string(successData)
-			}
-
-			complete := ""
-			if batchRequest.Complete != nil {
-				completeData, err := json.Marshal(batchRequest.Complete)
-				if err != nil {
-					_ = c.Error(cmd, fmt.Errorf("Invalid Complete job"))
-					return
-				}
-				complete = string(completeData)
-			}
-
-			meta := b.newBatchMeta(batchRequest.Description, success, complete)
-			batch, err := b.newBatch(batchId, meta)
-
-			if err != nil {
-				c.Error(cmd, fmt.Errorf("Unable to create batch: %v", err))
-				return
-			}
-
-			_ = c.Result([]byte(batch.Id))
-			return
-		case "OPEN":
-			batchId := parts[1]
-
-			// worker ids are usually only associated with workers
-			// in order for a client to submit a job to a batch it must pass wid to the payload when submitting HELO
-
-			// to retrieve the worker id for this request
-			// we must access client which is a private field of Connection
-			// use reflection in order to get the worker id that is requesting to open the batch
-			connection := reflect.ValueOf(*c)
-			client := connection.FieldByName("client").Elem()
-
-			wid := client.FieldByName("Wid").String()
-			if wid == "" {
-				_ = c.Error(cmd, fmt.Errorf("Batches can only be opened from a client with wid set"))
-				return
-			}
-
-			batch, err := b.getBatch(batchId)
-			if err != nil {
-				_ = c.Error(cmd, fmt.Errorf("Cannot get batch: %v", err))
-				return
-			}
-
-			if batch.isBatchDone() {
-				_ = c.Error(cmd, errors.New("Batch has already finished"))
-				return
-			}
-
-			if !batch.hasWorker(wid) {
-				_ = c.Error(cmd, fmt.Errorf("This worker is not working on a job in the requested batch"))
-				return
-			}
-
-			if err := batch.open(); err != nil {
-				_ = c.Error(cmd, fmt.Errorf("Cannot open batch: %v", err))
-				return
-			}
-
-			_ = c.Result([]byte(batch.Id))
-			return
-		case "COMMIT":
-			batchId := parts[1]
-			if batchId == "" {
-				_ = c.Error(cmd, errors.New("bid is required"))
-				return
-			}
-			batch, err := b.getBatch(batchId)
-			if err != nil {
-				_ = c.Error(cmd, fmt.Errorf("Cannot get batch: %v", err))
-				return
-			}
-
-			if err := batch.commit(); err != nil {
-				_ = c.Error(cmd, fmt.Errorf("Cannot commit batch: %v", err))
-				return
-			}
-			_ = c.Ok()
-			return
-		case "STATUS":
-			batchId := parts[1]
-			batch, err := b.getBatch(batchId)
-			if err != nil {
-				_ = c.Error(cmd, fmt.Errorf("Cannot find batch: %v", err))
-				return
-			}
-			data, err := json.Marshal(map[string]interface{}{
-				"bid":          batchId,
-				"total":        batch.Meta.Total,
-				"pending":      batch.Meta.Total - batch.Meta.Failed - batch.Meta.Succeeded,
-				"description":  batch.Meta.Description,
-				"created_at":   batch.Meta.CreatedAt,
-				"completed_st": CallbackJobPending,
-				"success_st":   CallbackJobPending,
-			})
-			if err != nil {
-				c.Error(cmd, fmt.Errorf("Unable to marshal batch data: %v", err))
-				return
-			}
-			_ = c.Result([]byte(data))
-			return
-		default:
-			_ = c.Error(cmd, fmt.Errorf("Invalid BATCH operation %s", parts[0]))
-		}
-
-	}
 }
 
 func (b *BatchSubsystem) addMiddleware() {
@@ -226,104 +79,6 @@ func (b *BatchSubsystem) getBatchFromInterface(batchId interface{}) (*batch, err
 		return nil, fmt.Errorf("Unable to get batch: %s", bid)
 	}
 	return batch, nil
-}
-
-func (b *BatchSubsystem) pushMiddleware(next func() error, ctx manager.Context) error {
-	if bid, ok := ctx.Job().GetCustom("bid"); ok {
-		batch, err := b.getBatchFromInterface(bid)
-		if err != nil {
-			return fmt.Errorf("Unable to get batch %s", bid)
-		}
-		if err := batch.jobQueued(ctx.Job().Jid); err != nil {
-			util.Warnf("Unable to add batch %v", err)
-			return fmt.Errorf("Unable to add job %s to batch %s", ctx.Job().Jid, bid)
-		}
-		util.Infof("Added %s to batch %s", ctx.Job().Jid, batch.Id)
-	}
-	return next()
-}
-
-func (b *BatchSubsystem) Fetch(ctx context.Context, wid string, queues ...string) (manager.Lease, error) {
-	lease, err := b.Fetcher.Fetch(ctx, wid, queues...)
-	if err == nil && lease != manager.Nothing {
-		job, err := lease.Job()
-		if err == nil && job != nil {
-			if bid, ok := job.GetCustom("bid"); ok {
-				batch, err := b.getBatchFromInterface(bid)
-				if err != nil {
-					return nil, fmt.Errorf("Unable to retrieve batch %s", bid)
-				}
-				batch.setWorkerForJid(job.Jid, wid)
-				util.Infof("Added worker %s for job %s to %s", wid, job.Jid, batch.Id)
-			}
-		}
-
-	}
-	return lease, err
-}
-
-func (b *BatchSubsystem) fetchMiddleware(next func() error, ctx manager.Context) error {
-	middleware_err := next() // runs the rest of the middleware
-	if bid, ok := ctx.Job().GetCustom("bid"); ok {
-		batch, err := b.getBatchFromInterface(bid)
-		if err != nil {
-			return fmt.Errorf("Unable to retrieve batch %s", bid)
-		}
-		if middleware_err != nil {
-			// clear the worker id for a job here since we set the worker id in the custom fetcher
-			batch.removeWorkerForJid(ctx.Job().Jid)
-		}
-	}
-
-	return middleware_err
-}
-
-func (b *BatchSubsystem) handleJobFinished(success bool) func(next func() error, ctx manager.Context) error {
-	return func(next func() error, ctx manager.Context) error {
-		if success {
-			// check if this is a success / complete job from batch
-			if bid, ok := ctx.Job().GetCustom("_bid"); ok {
-				batch, err := b.getBatchFromInterface(bid)
-				if err != nil {
-					util.Warnf("Unable to retrieve batch %s: %v", bid, err)
-					return next()
-				}
-				cb, ok := ctx.Job().GetCustom("_cb")
-				if !ok {
-					util.Warnf("Batch (%s) callback job (%s) does not have _cb specified", bid, ctx.Job().Type)
-					return next()
-				}
-				callbackType, ok := cb.(string)
-				if !ok {
-					util.Warnf("Error converting callback job type %s", cb)
-					return next()
-				}
-				if err := batch.callbackJobSucceded(callbackType); err != nil {
-					util.Warnf("Unable to update batch")
-				}
-				return next()
-			}
-		}
-		if bid, ok := ctx.Job().GetCustom("bid"); ok {
-			batch, err := b.getBatchFromInterface(bid)
-			if err != nil {
-				return fmt.Errorf("Unable to retrieve batch %s", bid)
-			}
-
-			status := "succeeded"
-			if !success {
-				status = "failed"
-			}
-			util.Infof("Job %s (worker %s) %s for batch %s", ctx.Job().Jid, ctx.Reservation().Wid, status, batch.Id)
-			batch.removeWorkerForJid(ctx.Job().Jid)
-			if err := batch.jobFinished(ctx.Job().Jid, success); err != nil {
-				util.Warnf("error processing finished job for batch %v", err)
-				return fmt.Errorf("Unable to process finished job %s for batch %s", ctx.Job().Jid, batch.Id)
-			}
-
-		}
-		return next()
-	}
 }
 
 func (b *BatchSubsystem) loadExistingBatches() error {
